Remove useless account API wrappers
