+++ src/actions/SpendingLimitsActions.js
@@ -4,7 +4,6 @@ import { Alert } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 
 import s from '../locales/strings.js'
-import { checkPassword } from '../modules/Core/Account/api.js'
 import * as SETTINGS_API from '../modules/Core/Account/settings.js'
 import { getAccount } from '../modules/Core/selectors.js'
 import { newSpendingLimits } from '../reducers/SpendingLimitsReducer.js'
@@ -15,14 +14,12 @@ export const setSpendingLimits = (spendingLimits: SpendingLimits, password: stri
   const state = getState()
   const account = getAccount(state)
 
-  Promise.resolve()
-    .then(() => checkPassword(account, password))
-    .then(isAuthorized => {
-      if (!isAuthorized) return Alert.alert(s.strings.password_check_incorrect_password_title)
+  account.checkPassword(password).then(isAuthorized => {
+    if (!isAuthorized) return Alert.alert(s.strings.password_check_incorrect_password_title)
 
-      return SETTINGS_API.setSpendingLimits(account, spendingLimits).then(() => {
-        dispatch(newSpendingLimits(spendingLimits))
-        Actions.pop()
-      })
+    return SETTINGS_API.setSpendingLimits(account, spendingLimits).then(() => {
+      dispatch(newSpendingLimits(spendingLimits))
+      Actions.pop()
     })
+  })
 }
